refactor: consolidate more integration tests (fastembed, working-memory-additive, working-memory-processor)
