+++ packages/memory/integration-tests/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "version": "0.1.0",
   "scripts": {
-    "test": "pnpm compile:worker && pnpm test:pg && pnpm test:upstash && pnpm test:libsql && pnpm test:streaming && pnpm test:working-memory && pnpm test:agent-memory",
+    "test": "pnpm compile:worker && pnpm test:pg && pnpm test:upstash && pnpm test:libsql && pnpm test:streaming && pnpm test:working-memory && pnpm test:agent-memory && pnpm test:fastembed",
     "pretest:pg": "docker compose up -d postgres && (for i in $(seq 1 30); do docker compose exec -T postgres pg_isready -U postgres && break || (sleep 1; [ $i -eq 30 ] && exit 1); done)",
     "test:pg": "vitest run ./src/with-pg-storage.test.ts",
     "posttest:pg": "docker compose down --volumes",
@@ -13,7 +13,7 @@
     "test:agent-memory": "vitest run ./src/agent-memory.test.ts",
     "test:libsql": "vitest run ./src/with-libsql-storage.test.ts",
     "test:streaming": "vitest run ./src/streaming-memory.test.ts",
-    "test:working-memory": "vitest run ./src/working-memory.test.ts",
+    "test:working-memory": "vitest run ./src/working-memory.test.ts && vitest run ./src/working-memory-processor.test.ts",
     "pretest:pg:perf": "docker compose up -d postgres && (for i in $(seq 1 30); do docker compose exec -T postgres pg_isready -U postgres && break || (sleep 1; [ $i -eq 30 ] && exit 1); done)",
     "test:processors": "echo 'processors have moved into core. we can delete this command after the pr it came in on is merged.'",
     "test:pg:perf": "vitest run ./src/performance-testing/with-pg-storage.test.ts",
@@ -25,6 +25,7 @@
     "test:libsql:perf": "vitest run ./src/performance-testing/with-libsql-storage.test.ts",
     "posttest:libsql:perf": "docker compose down --volumes",
     "test:perf": "pnpm test:pg:perf && pnpm test:upstash:perf && pnpm test:libsql:perf",
+    "test:fastembed": "vitest run ./src/fastembed.test.ts",
     "compile:worker": "npx tsc ./src/worker/generic-memory-worker.ts ./src/worker/mock-embedder.ts --esModuleInterop --resolveJsonModule --module commonjs --target es2020 --outDir ./ --rootDir ./ --skipLibCheck",
     "dev": "mastra dev"
   },
@@ -51,6 +52,7 @@
     "@types/json-schema": "^7.0.15",
     "@types/node": "22.13.17",
     "ai": "^4.2.2",
+    "ai-v5": "npm:ai@^5.0.0",
     "ai-v6": "npm:ai@latest",
     "jsdom": "^26.1.0",
     "mastra": "*",
