+++ packages/memory/integration-tests/src/working-memory.test.ts
@@ -2,6 +2,7 @@ import { openai } from '@ai-sdk/openai';
 import { openai as openaiV6 } from '@ai-sdk/openai-v6';
 import { config } from 'dotenv';
 import { getWorkingMemoryTests } from './shared/working-memory';
+import { getWorkingMemoryAdditiveTests } from './shared/working-memory-additive';
 
 config({ path: '.env.test' });
 
@@ -9,13 +10,16 @@ config({ path: '.env.test' });
 getWorkingMemoryTests({
   model: openai('gpt-4o'),
 });
+getWorkingMemoryAdditiveTests(openai('gpt-4o'));
 
 // v5
 getWorkingMemoryTests({
   model: 'openai/gpt-4o',
 });
+getWorkingMemoryAdditiveTests('openai/gpt-4o');
 
 // v6
 getWorkingMemoryTests({
   model: openaiV6('gpt-4o'),
 });
+getWorkingMemoryAdditiveTests(openaiV6('gpt-4o'));
